+++ bb/tasks.clj
@@ -35,7 +35,7 @@
 (defn publish []
   (build-squint-npm-package)
   (run! fs/delete (fs/glob "lib" "*.map"))
-  (shell "esbuild src/squint/core.js --minify --format=iife --global-name=squint.core --outfile=lib/squint.core.umd.js")
+  (shell "npx esbuild src/squint/core.js --minify --format=iife --global-name=squint.core --outfile=lib/squint.core.umd.js")
   (shell "npm publish"))
 
 (defn watch-squint []